Simple code for reading temperature/pH. Not tested, not calibrated.
The file sensors.ino reads the analog input (10 bit number == 1024) from the sensors,
then, turns into a voltage value. Finally, the voltage is turn into a usefull value,
depending on the sensor, and is printed to the user.

Note: The formula for converting a voltage to a usefull value is dependent on the sensor
you are using. Check the vendors datasheet for more information about the sensor
Note2: The temperature and pH sensor are from atlas-scientific.

The file sketch.yaml stores configuration information for the Arduino CLI. This file is
not required but makes the Arduino CLI easier to use.